@@ -65,9 +65,9 @@
   },
   "homepage": "https://github.com/Semantic-Org/Semantic-UI-React#readme",
   "dependencies": {
-    "@babel/runtime": "^7.10.5",
+    "@babel/runtime": "^7.23.7",
     "@fluentui/react-component-event-listener": "~0.63.0",
-    "@popperjs/core": "^2.6.0",
+    "@popperjs/core": "^2.11.8",
     "@semantic-ui-react/event-stack": "^3.1.3",
     "clsx": "^1.1.1",
     "keyboard-key": "^1.1.0",
@@ -80,17 +80,14 @@
   },
   "devDependencies": {
     "@artsy/fresnel": "^1.2.0",
-    "@babel/cli": "^7.10.5",
-    "@babel/core": "^7.10.5",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-export-default-from": "^7.10.4",
-    "@babel/plugin-proposal-export-namespace-from": "^7.10.4",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.10.5",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/register": "^7.10.5",
-    "@babel/standalone": "^7.10.5",
+    "@babel/cli": "^7.23.4",
+    "@babel/core": "^7.23.7",
+    "@babel/plugin-proposal-export-default-from": "^7.23.3",
+    "@babel/plugin-transform-runtime": "^7.23.7",
+    "@babel/preset-env": "^7.23.7",
+    "@babel/preset-react": "^7.23.3",
+    "@babel/register": "^7.23.7",
+    "@babel/standalone": "^7.23.7",
     "@mdx-js/loader": "^0.20.3",
     "@percy/cli": "^1.0.0-beta.73",
     "@percy/cypress": "^3.1.1",
@@ -103,7 +100,7 @@
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "babel-plugin-filter-imports": "^4.0.0",
-    "babel-plugin-istanbul": "^6.0.0",
+    "babel-plugin-istanbul": "^6.1.1",
     "babel-plugin-lodash": "^3.3.4",
     "babel-plugin-transform-react-handled-props": "^2.1.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
@@ -152,7 +149,7 @@
     "react": "^17.0.0",
     "react-ace": "^6.4.0",
     "react-codesandboxer": "^3.1.3",
-    "react-docgen": "^4.1.0",
+    "react-docgen": "^7.0.1",
     "react-dom": "^17.0.0",
     "react-hot-loader": "^4.13.0",
     "react-intersection-observer": "^8.26.2",
@@ -175,7 +172,7 @@
     "terser-webpack-plugin": "^3.0.8",
     "terser-webpack-plugin-legacy": "^1.2.3",
     "through2": "^3.0.1",
-    "typescript": "^3.9.5",
+    "typescript": "^4.5.5",
     "vinyl": "^2.2.0",
     "webpack": "^4.28.4",
     "webpack-bundle-analyzer": "^3.8.0",
